api/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.2
sqlalchemy==2.0.41
pydantic==2.11.5
pydantic-settings==2.9.1
PyJWT==2.10.1
httpx==0.28.1
beautifulsoup4==4.13.4
python-multipart==0.0.20
pytest==8.3.5
